requirements.txt
python-dotenv
yfinance
plotly
rich
pyyaml
streamlit

# LLMs & Frameworks
llama-index
agno
google-generativeai

